refactor(tui): move thinking history renderer
Extract reasoning/thinking summary extraction, rendering, glyph constants, and color-depth cache from history.rs into history/thinking.rs while preserving the existing history API and test surface.

Refs #3308.
